fix: Use SASS modules to resolve deprecation warnings (#648)
* fix: updated Grid Sass List.nth

* fix: Use SASS modules to resolve deprecation warnings

* fix: rebuilt icons

* fix: sass lightness and @use "sass:color";

* fix: Restore URL escaping from previous commit.

It was unclear what this escaping was for, but now it is apparently it is for
escaping `#` at the start of HTML colours.  Previously this was applied to the
whole string, however this change restores it only escape the incoming colour
values rather than the whole string.

* chore: Move icon SASS from JS into an .scss file, and use tsx instead of tsc.

This makes it easier to syntax check, and also allows running the TS code
directly, instead of having to build it first.  This makes the build process
much cleaner as it removes the TS transpilation step which sometimes would
cause old code to run after making changes.

* linting

* updated package-lock

* fix: Use SASS modules to resolve deprecation warnings

* fix: rebuilt icons

* fix: Restore URL escaping from previous commit.

It was unclear what this escaping was for, but now it is apparently it is for
escaping `#` at the start of HTML colours.  Previously this was applied to the
whole string, however this change restores it only escape the incoming colour
values rather than the whole string.

* chore: Move icon SASS from JS into an .scss file, and use tsx instead of tsc.

This makes it easier to syntax check, and also allows running the TS code
directly, instead of having to build it first.  This makes the build process
much cleaner as it removes the TS transpilation step which sometimes would
cause old code to run after making changes.

* updated package-lock

* fix: linter

* fix: Deprecation warning

* updated map.get

* fix: Sass deprecations fixes

* Update Sass deprecations

* updated style references

* fix: Use SASS modules to resolve deprecation warnings

* fix: rebuilt icons

* fix: Restore URL escaping from previous commit.

It was unclear what this escaping was for, but now it is apparently it is for
escaping `#` at the start of HTML colours.  Previously this was applied to the
whole string, however this change restores it only escape the incoming colour
values rather than the whole string.

* chore: Move icon SASS from JS into an .scss file, and use tsx instead of tsc.

This makes it easier to syntax check, and also allows running the TS code
directly, instead of having to build it first.  This makes the build process
much cleaner as it removes the TS transpilation step which sometimes would
cause old code to run after making changes.

* updated package-lock

* fix: updated Grid Sass List.nth

* fix: sass lightness and @use "sass:color";

* linting

* fix: rebuilt icons

* fix: linter

* updated map.get

* fix: Deprecation warning

* fix: Sass deprecations fixes

* Update Sass deprecations

* updated style references

* icon update

---------

Co-authored-by: uqbcrunk <b.crunkhorn@uq.edu.au>
Co-authored-by: Brett <github@crunkhorn.com.au>